RF: move _wipe_out_offdiag to the test module
